Переписывает код проекта из файла pictures(закомментирован) в стиле AMD: добавляет модули picture, load, filter, filter-type, utilities и подключает их как зависимости в файл main.js.
