+++ src/filter/filter-type.js
@@ -1,0 +1,15 @@
+/**
+ * @fileoverview Список доступных видов фильтрации
+ * @author Oksana Semonenko
+ */
+
+'use strict';
+
+/** @enum {number} */
+define(function() {
+  return {
+    'POPULAR': 'popular',
+    'NEW': 'new',
+    'DISCUSSED': 'discussed'
+  };
+});
